HomeEstate/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.1.5
django-cors-headers==4.6.0
django-environ==0.12.0
django-rest-knox==5.0.2
djangorestframework==3.15.2
idna==3.10
pillow==11.1.0
psycopg2==2.9.10
python-barcode==0.15.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
